+++ sphynx/python/derive_html.py
@@ -1,7 +1,5 @@
 '''Run user code.'''
 import io
-import base64
-from matplotlib import pyplot as plt
 import numpy as np
 import pandas as pd
 import os
@@ -45,7 +43,34 @@ except BaseException:
   traceback.print_exception(a, b, c.tb_next)
   sys.exit(1)
 
-f = io.BytesIO()
-plt.savefig(f, format='svg')
-data = base64.b64encode(f.getvalue()).decode('ascii')
-op.output_scalar('sc', f'<img src="data:image/svg+xml;base64, {data}">')
+if op.params['mode'] == 'html':
+  assert 'html' in globals(), 'Please save the output as "html".'
+  sc = html
+elif op.params['mode'] == 'matplotlib':
+  import base64
+  from matplotlib import pyplot as plt
+  f = io.BytesIO()
+  plt.savefig(f, format='svg')
+  data = base64.b64encode(f.getvalue()).decode('ascii')
+  sc = f'<img src="data:image/svg+xml;base64, {data}">'
+elif op.params['mode'] == 'ipywidget':
+  from ipywidgets.embed import embed_minimal_html
+  f = io.StringIO()
+  embed_minimal_html(f, views=None)
+  sc = f.getvalue()
+  # The widget code has to be loaded from a CDN.
+  sc = sc.replace('embed-amd.js"', 'embed-amd.js" data-jupyter-widgets-cdn-only')
+  # We patch the URL class because this is displayed via <iframe srcdoc="...">.
+  sc = sc.replace('<body>', '''
+    <body>
+    <script>
+      const RURL = URL;
+    </script>
+    <script>
+      class URL extends RURL {
+        constructor(url, base) {
+          super(url, 'http://example.com/');
+        }
+      }
+    </script>''')
+op.output_scalar('sc', sc)